Move kernel telemetry into kernel folder
