+++ src/kernels/execution/kernelExecution.ts
@@ -12,7 +12,8 @@ import { traceInfo, traceInfoIfCI, traceWarning } from '../../platform/logging';
 import { IDisposable, IExtensionContext } from '../../platform/common/types';
 import { createDeferred, waitForPromise } from '../../platform/common/utils/async';
 import { StopWatch } from '../../platform/common/utils/stopWatch';
-import { trackKernelResourceInformation, sendKernelTelemetryEvent } from '../../telemetry/telemetry';
+import { sendKernelTelemetryEvent } from '../telemetry/sendKernelTelemetryEvent';
+import { trackKernelResourceInformation } from '../telemetry/helper';
 import { captureTelemetry } from '../../telemetry';
 import { Telemetry } from '../../webviews/webview-side/common/constants';
 import { CellOutputDisplayIdTracker } from './cellDisplayIdTracker';
