+++ src/kernels/helpers.node.ts
@@ -10,7 +10,7 @@ import { traceError, traceVerbose } from '../platform/logging';
 import { getDisplayPath } from '../platform/common/platform/fs-paths';
 import { IPythonExecutionFactory } from '../platform/common/process/types.node';
 import { Resource } from '../platform/common/types';
-import { trackKernelResourceInformation } from '../telemetry/telemetry';
+import { trackKernelResourceInformation } from './telemetry/helper';
 import { areInterpreterPathsSame } from '../platform/pythonEnvironments/info/interpreter';
 import { sendTelemetryEvent } from '../telemetry';
 import { concatMultilineString } from '../webviews/webview-side/common';
