+++ src/kernels/jupyter/launcher/liveshare/hostJupyterServer.ts
@@ -20,7 +20,7 @@ import { createDeferred, sleep } from '../../../../platform/common/utils/async';
 import { DataScience } from '../../../../platform/common/utils/localize';
 import { StopWatch } from '../../../../platform/common/utils/stopWatch';
 import { SessionDisposedError } from '../../../../platform/errors/sessionDisposedError';
-import { sendKernelTelemetryEvent } from '../../../../telemetry/telemetry';
+import { sendKernelTelemetryEvent } from '../../../telemetry/sendKernelTelemetryEvent';
 import { Telemetry } from '../../../../webviews/webview-side/common/constants';
 import {
     KernelConnectionMetadata,
