+++ src/kernels/raw/session/hostRawNotebookProvider.node.ts
@@ -19,7 +19,7 @@ import {
 } from '../../../platform/common/types';
 import { createDeferred } from '../../../platform/common/utils/async';
 import { DataScience } from '../../../platform/common/utils/localize';
-import { trackKernelResourceInformation } from '../../../telemetry/telemetry';
+import { trackKernelResourceInformation } from '../../telemetry/helper';
 import { captureTelemetry, sendTelemetryEvent } from '../../../telemetry';
 import { Telemetry } from '../../../webviews/webview-side/common/constants';
 import { isPythonKernelConnection } from '../../helpers';
