+++ src/notebooks/controllers/notebookControllerManager.ts
@@ -35,7 +35,7 @@ import {
 import { noop } from '../../platform/common/utils/misc';
 import { StopWatch } from '../../platform/common/utils/stopWatch';
 import { sendKernelListTelemetry } from '../telemetry/kernelTelemetry';
-import { trackKernelResourceInformation } from '../../telemetry/telemetry';
+import { trackKernelResourceInformation } from '../../kernels/telemetry/helper';
 import { IInterpreterService } from '../../platform/interpreter/contracts';
 import { IServiceContainer } from '../../platform/ioc/types';
 import { PythonEnvironment } from '../../platform/pythonEnvironments/info';
