+++ src/telemetry/telemetry.ts
@@ -4,53 +4,20 @@
 // eslint-disable-next-line @typescript-eslint/no-require-imports
 import cloneDeep = require('lodash/cloneDeep');
 import { Uri } from 'vscode';
-import { InterruptResult, KernelConnectionMetadata } from '../kernels/types';
 import { Resource } from '../platform/common/types';
-import { StopWatch } from '../platform/common/utils/stopWatch';
 import { IInterpreterPackages, ResourceSpecificTelemetryProperties } from './types';
-import { WorkspaceInterpreterTracker } from './workspaceInterpreterTracker';
-import { PYTHON_LANGUAGE, Telemetry } from '../platform/common/constants';
-import { InterpreterCountTracker } from './interpreterCountTracker';
-import { getTelemetrySafeHashedString, getTelemetrySafeLanguage } from './helpers';
+import { getTelemetrySafeHashedString } from './helpers';
 import { PythonEnvironment } from '../platform/pythonEnvironments/info';
 import { createDeferred } from '../platform/common/utils/async';
-import { getNormalizedInterpreterPath } from '../platform/pythonEnvironments/info/interpreter';
 import { getResourceType } from '../platform/common/utils';
-import { populateTelemetryWithErrorInfo } from '../platform/errors';
-import { setSharedProperty, IEventNamePropertyMapping, sendTelemetryEvent, waitBeforeSending } from '.';
 import { IServiceContainer } from '../platform/ioc/types';
 import { getComparisonKey } from '../platform/vscode-path/resources';
-import { getFilePath } from '../platform/common/platform/fs-paths';
-
-/**
- * This information is sent with each telemetry event.
- */
-type ContextualTelemetryProps = {
-    /**
-     * Whether we're starting the preferred kernel or not.
-     * If false, then the user chose a different kernel when starting the notebook.
-     * Doesn't really apply to Interactive Window, as we always pick the current interpreter.
-     */
-    isPreferredKernel?: boolean;
-    kernelConnection: KernelConnectionMetadata;
-    startFailed: boolean;
-    kernelDied: boolean;
-    interruptKernel: boolean;
-    restartKernel: boolean;
-    kernelSpecCount: number; // Total number of kernel specs in list of kernels.
-    kernelInterpreterCount: number; // Total number of interpreters in list of kernels
-    kernelLiveCount: number; // Total number of live kernels in list of kernels.
-    /**
-     * When we start local Python kernels, this property indicates whether the interpreter matches the kernel. If not this means we've started the wrong interpreter or the mapping is wrong.
-     */
-    interpreterMatchesKernel: boolean;
-};
 
 type Context = {
     previouslySelectedKernelConnectionId: string;
 };
-const trackedInfo = new Map<string, [ResourceSpecificTelemetryProperties, Context]>();
-const pythonEnvironmentsByHash = new Map<string, PythonEnvironment>();
+export const trackedInfo = new Map<string, [ResourceSpecificTelemetryProperties, Context]>();
+export const pythonEnvironmentsByHash = new Map<string, PythonEnvironment>();
 let globalContainer: IServiceContainer | undefined = undefined;
 
 export function initializeGlobals(serviceContainer: IServiceContainer) {
@@ -58,213 +25,11 @@ export function initializeGlobals(serviceContainer: IServiceContainer) {
 }
 
 /**
- * Initializes the Interactive/Notebook telemetry as a result of user action.
- */
-export function initializeInteractiveOrNotebookTelemetryBasedOnUserAction(
-    resourceUri: Resource,
-    kernelConnection: KernelConnectionMetadata
-) {
-    setSharedProperty('userExecutedCell', 'true');
-    trackKernelResourceInformation(resourceUri, { kernelConnection });
-}
-/**
- * @param {(P[E] & { waitBeforeSending: Promise<void> })} [properties]
- * Can optionally contain a property `waitBeforeSending` referencing a promise.
- * Which must be awaited before sending the telemetry.
- */
-export function sendKernelTelemetryEvent<P extends IEventNamePropertyMapping, E extends keyof P>(
-    resource: Resource,
-    eventName: E,
-    durationMs?: Record<string, number> | number,
-    properties?: P[E] & { waitBeforeSending?: Promise<void> },
-    ex?: Error
-) {
-    if (eventName === Telemetry.ExecuteCell) {
-        setSharedProperty('userExecutedCell', 'true');
-    }
-
-    const props = getContextualPropsForTelemetry(resource);
-    Object.assign(props, properties || {});
-    sendTelemetryEvent(
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        eventName as any,
-        durationMs,
-        props,
-        ex,
-        true
-    );
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    resetData(resource, eventName as any, props);
-
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    incrementStartFailureCount(resource, eventName as any, props);
-}
-
-/**
- * Send this & subsequent telemetry only after this promise has been resolved.
- * We have a default timeout of 30s.
- * @param {P[E]} [properties]
- * Can optionally contain a property `waitBeforeSending` referencing a promise.
- * Which must be awaited before sending the telemetry.
- */
-export function sendKernelTelemetryWhenDone<P extends IEventNamePropertyMapping, E extends keyof P>(
-    resource: Resource,
-    eventName: E,
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    promise: Promise<any> | Thenable<any>,
-    stopWatch?: StopWatch,
-    properties?: P[E] & { [waitBeforeSending]?: Promise<void> }
-) {
-    if (eventName === Telemetry.ExecuteCell) {
-        setSharedProperty('userExecutedCell', 'true');
-    }
-    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    const props: any = properties || {};
-    stopWatch = stopWatch ? stopWatch : new StopWatch();
-    if (typeof promise.then === 'function') {
-        // eslint-disable-next-line , @typescript-eslint/no-explicit-any
-        (promise as Promise<any>)
-            .then(
-                (data) => {
-                    const props = getContextualPropsForTelemetry(resource);
-                    Object.assign(props, properties || {});
-                    sendTelemetryEvent(
-                        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                        eventName as any,
-                        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                        stopWatch!.elapsedTime,
-                        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                        props as any
-                    );
-                    return data;
-                    // eslint-disable-next-line @typescript-eslint/promise-function-async
-                },
-                (ex) => {
-                    const props = getContextualPropsForTelemetry(resource);
-                    Object.assign(props, properties || {});
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    populateTelemetryWithErrorInfo(props as any, ex);
-                    sendTelemetryEvent(
-                        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                        eventName as any,
-                        // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
-                        stopWatch!.elapsedTime,
-                        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                        props as any,
-                        ex,
-                        true
-                    );
-                    // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                    incrementStartFailureCount(resource, eventName as any, props);
-                }
-            )
-            .finally(() => {
-                // eslint-disable-next-line @typescript-eslint/no-explicit-any
-                resetData(resource, eventName as any, props);
-            });
-    }
-}
-export function trackKernelResourceInformation(resource: Resource, information: Partial<ContextualTelemetryProps>) {
-    if (!resource) {
-        return;
-    }
-    const key = getComparisonKey(resource);
-    const [currentData, context] = trackedInfo.get(key) || [
-        {
-            resourceType: getResourceType(resource),
-            resourceHash: resource ? getTelemetrySafeHashedString(resource.toString()) : undefined,
-            kernelSessionId: getTelemetrySafeHashedString(Date.now().toString())
-        },
-        { previouslySelectedKernelConnectionId: '' }
-    ];
-
-    if (information.restartKernel) {
-        currentData.kernelSessionId = getTelemetrySafeHashedString(Date.now().toString());
-        currentData.interruptCount = 0;
-        currentData.restartCount = (currentData.restartCount || 0) + 1;
-    }
-    if (information.interruptKernel) {
-        currentData.interruptCount = (currentData.interruptCount || 0) + 1;
-    }
-    if (information.startFailed) {
-        currentData.startFailureCount = (currentData.startFailureCount || 0) + 1;
-    }
-    currentData.kernelSpecCount = information.kernelSpecCount || currentData.kernelSpecCount || 0;
-    currentData.kernelLiveCount = information.kernelLiveCount || currentData.kernelLiveCount || 0;
-    currentData.kernelInterpreterCount = information.kernelInterpreterCount || currentData.kernelInterpreterCount || 0;
-    currentData.pythonEnvironmentCount = InterpreterCountTracker.totalNumberOfInterpreters;
-
-    const kernelConnection = information.kernelConnection;
-    if (kernelConnection) {
-        const newKernelConnectionId = kernelConnection.id;
-        // If we have selected a whole new kernel connection for this,
-        // Then reset some of the data
-        if (context.previouslySelectedKernelConnectionId !== newKernelConnectionId) {
-            clearInterruptCounter(resource);
-            clearRestartCounter(resource);
-        }
-        if (
-            context.previouslySelectedKernelConnectionId &&
-            context.previouslySelectedKernelConnectionId !== newKernelConnectionId
-        ) {
-            currentData.kernelSessionId = getTelemetrySafeHashedString(Date.now().toString());
-            currentData.switchKernelCount = (currentData.switchKernelCount || 0) + 1;
-        }
-        let language: string | undefined;
-        switch (kernelConnection.kind) {
-            case 'connectToLiveRemoteKernel':
-                language = kernelConnection.kernelModel.language;
-                break;
-            case 'startUsingRemoteKernelSpec':
-            case 'startUsingLocalKernelSpec':
-                language = kernelConnection.kernelSpec.language;
-                break;
-            case 'startUsingPythonInterpreter':
-                language = PYTHON_LANGUAGE;
-                break;
-            default:
-                break;
-        }
-        currentData.kernelLanguage = getTelemetrySafeLanguage(language);
-        // Keep track of the kernel that was last selected.
-        context.previouslySelectedKernelConnectionId = kernelConnection.id;
-
-        const interpreter = kernelConnection.interpreter;
-        if (interpreter) {
-            currentData.isUsingActiveInterpreter = WorkspaceInterpreterTracker.isActiveWorkspaceInterpreter(
-                resource,
-                interpreter
-            );
-            currentData.pythonEnvironmentType = interpreter.envType;
-            currentData.pythonEnvironmentPath = getTelemetrySafeHashedString(
-                getFilePath(getNormalizedInterpreterPath(interpreter.uri))
-            );
-            pythonEnvironmentsByHash.set(currentData.pythonEnvironmentPath, interpreter);
-            if (interpreter.version) {
-                const { major, minor, patch } = interpreter.version;
-                currentData.pythonEnvironmentVersion = `${major}.${minor}.${patch}`;
-            } else {
-                currentData.pythonEnvironmentVersion = undefined;
-            }
-
-            updatePythonPackages(currentData);
-        }
-
-        currentData.kernelConnectionType = currentData.kernelConnectionType || kernelConnection?.kind;
-    } else {
-        context.previouslySelectedKernelConnectionId = '';
-    }
-
-    trackedInfo.set(key, [currentData, context]);
-}
-
-/**
  * The python package information is fetch asynchronously.
  * Its possible the information is available at a later time.
  * Use this to update with the latest information (if available)
  */
-function updatePythonPackages(
+export function updatePythonPackages(
     currentData: ResourceSpecificTelemetryProperties & { waitBeforeSending?: Promise<void> },
     clonedCurrentData?: ResourceSpecificTelemetryProperties & {
         waitBeforeSending?: Promise<void>;
@@ -328,7 +93,7 @@ export function deleteTrackedInformation(resource: Uri) {
  * Always return a clone of the properties.
  * We will be using a reference of this object elsewhere & adding properties to the object.
  */
-function getContextualPropsForTelemetry(
+export function getContextualPropsForTelemetry(
     resource: Resource
 ): ResourceSpecificTelemetryProperties & { waitBeforeSendingTelemetry?: Promise<void> } {
     if (!resource) {
@@ -354,62 +119,4 @@ function getContextualPropsForTelemetry(
     // 2. Data that is returned & sent via telemetry now
     updatePythonPackages(data[0], clonedData);
     return clonedData;
-}
-/**
- * Some information such as interrupt counters & restart counters need to be reset
- * after we have successfully interrupted or restarted a kernel.
- */
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function resetData(resource: Resource, eventName: string, properties: any) {
-    // Once we have successfully interrupted, clear the interrupt counter.
-    if (eventName === Telemetry.NotebookInterrupt) {
-        let kv: Pick<IEventNamePropertyMapping, Telemetry.NotebookInterrupt>;
-        const data: undefined | typeof kv[Telemetry.NotebookInterrupt] = properties;
-        // Check result to determine if success.
-        if (data && 'result' in data && data.result === InterruptResult.Success) {
-            clearInterruptCounter(resource);
-        }
-    }
-    // Once we have successfully restarted, clear the interrupt counter.
-    if (eventName === Telemetry.NotebookRestart) {
-        let kv: Pick<IEventNamePropertyMapping, Telemetry.NotebookRestart>;
-        const data: undefined | typeof kv[Telemetry.NotebookRestart] = properties;
-        // For restart to be successful, we should not have `failed`
-        const failed = data && 'failed' in data ? data.failed : false;
-        if (!failed) {
-            clearInterruptCounter(resource);
-        }
-    }
-}
-function clearInterruptCounter(resource: Resource) {
-    if (!resource) {
-        return;
-    }
-    const key = getComparisonKey(resource);
-    const currentData = trackedInfo.get(key);
-    if (currentData) {
-        currentData[0].interruptCount = 0;
-    }
-}
-function clearRestartCounter(resource: Resource) {
-    if (!resource) {
-        return;
-    }
-    const key = getComparisonKey(resource);
-    const currentData = trackedInfo.get(key);
-    if (currentData) {
-        currentData[0].restartCount = 0;
-        currentData[0].startFailureCount = 0;
-    }
-}
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-function incrementStartFailureCount(resource: Resource, eventName: any, properties: any) {
-    if (eventName === Telemetry.NotebookStart) {
-        let kv: Pick<IEventNamePropertyMapping, Telemetry.NotebookStart>;
-        const data: undefined | typeof kv[Telemetry.NotebookStart] = properties;
-        // Check start failed.
-        if (data && 'failed' in data && data.failed) {
-            trackKernelResourceInformation(resource, { startFailed: true });
-        }
-    }
 }
